Yay! We have headline links on ads.index.php populating single ads on ads.show.php.
We needed this for a good user experience.
I also adjusted the H1 text on ads.show.php. Adusted slightly the Ads class.
CSS was adjusted to enlarge font, change color and center the font. Readability is helped, but would be assisted more by a left flush margin a couple inches in on the page. Perhaps use a %em so it is still good on mobile.