Yay! We are populating the database with the ads.create.php file.
Moving forward. We fixed the action part of the form tag to point to the ads.create.page.
This sent the data to the database through the code at the top of the file.
We are changing the ads.edit.php page to be an edit page after publishing of an ad.
We are going to add photo upload options to the ads.create.php page as an option.