Yay! Ads are creating and populating correctly.
We had trouble with the buttons, radial or checkbox type, in the form inputs. They would return a yes or no, but I could not figure out yet how to code them in the functions to populate correctly or nicely and also load in the database.
I removed all those variables, stripping down to the most basic of ads. I am sure that Craigslist started on much simpler terms than it is today, so it's okay.