Adjusted CSS in ads.index to display the ads centered on the page.
The unordered list tags were creating issues when trying to move the ads away from the left edge, and margin changes included the footer, so the footer would look really dumb. I adjusted the html tags for the ads only and created a paragraph instead of unordered list to ditch the unordered list bullets that wouldn't stick to the ads if the ads were centered, since margin changes were no possible with this theme.

Not perfectly readable, but good for starting out.