Started working the users create file. Had to adjust the forms so the data would populate.
Adjusted forms to match the database columns.
Adjusted form tags so that the data would populate on the database.
Added classes to form elements so I could CSS them.
Adjusted the CSS so that the form would be more centered on the page, not all the way left. Used percentage so it remains flexible for mobile I hope.